requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
firebase-admin==6.2.0
httpx==0.25.2
python-multipart==0.0.6
slowapi==0.1.9
pydantic==2.5.0